FF-33 #comment Separating out editable/form components into own JS file. ToDo: documentation, schema patterns
